requirements.txt
# Core dependencies
requests==2.31.0
PyGithub==2.8.1
pandas==2.3.3
openpyxl==3.1.5
deepdiff==8.6.1
PyYAML==6.0.1

# Dev dependencies
pytest==8.2.0
black==24.4.2
flake8==7.0.0
